NUMPY: Finish the exercises (with solution)
I skipped the basic ODE integration one, it seemed unnecessary.
Some of the questions are slightly changed, but it should be mostly
the same. Please feel free to go over this version, I will probably
not have a look at it until monday. It is considered final
(except possibly one more proofread)